optimize the base example
