--- examples/base/scripts/syncBatchRoot.js
@@ -1,109 +1,0 @@
-const base = require('@eth-optimism/sdk');
-const ethers = require('ethers');
-const { readDeployContract, getLogName } = require('../../../script/utils');
-const logName = require('../../../script/deploy_log_name');
-const { L1_MAINNET_CONTRACTS, L1_TESTNET_CONTRACTS } = require('./constants');
-const { task } = require('hardhat/config');
-require('dotenv').config();
-
-task('syncBatchRoot', 'Forward message to L2').setAction(async (taskArgs, hre) => {
-  const walletPrivateKey = process.env.DEVNET_PRIVKEY;
-  const l1Provider = new ethers.providers.StaticJsonRpcProvider(process.env.L1RPC);
-  const l2Provider = new ethers.providers.StaticJsonRpcProvider(process.env.L2RPC);
-  const ethereumName = process.env.ETHEREUM;
-  const baseName = process.env.BASE;
-  const l1Wallet = new ethers.Wallet(walletPrivateKey, l1Provider);
-  const l2Wallet = new ethers.Wallet(walletPrivateKey, l2Provider);
-
-  const messengerL1Contracts = ethereumName !== 'ETHEREUM' ? L1_TESTNET_CONTRACTS : L1_MAINNET_CONTRACTS;
-  const messenger = new base.CrossChainMessenger({
-    l1ChainId: await l1Wallet.getChainId(), // 11155111 for Sepolia, 1 for Ethereum
-    l2ChainId: await l2Wallet.getChainId(), // 84532 for Base Sepolia, 8453 for Base Mainnet
-    l1SignerOrProvider: l1Wallet,
-    l2SignerOrProvider: l2Wallet,
-    bedrock: true,
-    contracts: {
-      l1: messengerL1Contracts,
-    },
-  });
-
-  const l1WalletAddress = await l1Wallet.getAddress();
-  const l1WalletBalance = ethers.utils.formatEther(await l1Wallet.getBalance());
-  console.log(`${l1WalletAddress} balance on l1: ${l1WalletBalance} ether`);
-
-  const arbitratorAddr = readDeployContract(
-    logName.DEPLOY_ARBITRATOR_LOG_PREFIX,
-    logName.DEPLOY_LOG_ARBITRATOR,
-    ethereumName,
-  );
-  if (arbitratorAddr === undefined) {
-    console.log('The arbitrator address not exist');
-    return;
-  }
-  console.log(`The arbitrator address: ${arbitratorAddr}`);
-
-  const zkLinkAddr = readDeployContract(logName.DEPLOY_ZKLINK_LOG_PREFIX, logName.DEPLOY_LOG_ZKLINK_PROXY, baseName);
-  if (zkLinkAddr === undefined) {
-    console.log('zkLink address not exist');
-    return;
-  }
-  console.log(`The zkLink address: ${zkLinkAddr}`);
-
-  const l1GatewayLogName = getLogName(logName.DEPLOY_L1_GATEWAY_LOG_PREFIX, baseName);
-  const baseL1GatewayAddr = readDeployContract(l1GatewayLogName, logName.DEPLOY_GATEWAY, ethereumName);
-  if (baseL1GatewayAddr === undefined) {
-    console.log('base l1 gateway address not exist');
-    return;
-  }
-  console.log(`The base l1 gateway address: ${baseL1GatewayAddr}`);
-
-  const baseL2GatewayAddr = readDeployContract(logName.DEPLOY_L2_GATEWAY_LOG_PREFIX, logName.DEPLOY_GATEWAY, baseName);
-  if (baseL2GatewayAddr === undefined) {
-    console.log('base l2 gateway address not exist');
-    return;
-  }
-  console.log(`The base l2 gateway address: ${baseL2GatewayAddr}`);
-
-  // pre-execution calldata
-  const zkLink = await hre.ethers.getContractAt('DummyZkLink', zkLinkAddr, l2Wallet);
-  const zklinkIface = zkLink.interface;
-  const blockNumber = await l2Provider.getBlockNumber();
-  console.log(`The current block number: ${blockNumber}`);
-  const l2LogsRootHash = ethers.utils.keccak256(ethers.utils.toUtf8Bytes(`L2 logs root hash ${blockNumber}`));
-  console.log(`The l2 logs root hash: ${l2LogsRootHash}`);
-  const executeCalldata = zklinkIface.encodeFunctionData('syncBatchRoot', [blockNumber, l2LogsRootHash, 0]);
-  console.log(`The call data: ${executeCalldata}`);
-  const gateway = await hre.ethers.getContractAt('OptimismGateway', baseL2GatewayAddr, l2Wallet);
-  const sendData = gateway.interface.encodeFunctionData('claimMessageCallback', [0, executeCalldata]);
-
-  const gasLimit = await messenger.estimateGas.sendMessage({
-    direction: 1, // L2_TO_L1, Estimating the Gas Required on L2
-    target: baseL2GatewayAddr,
-    message: sendData,
-  });
-  console.log(`The gas limit: ${gasLimit}`);
-
-  // forward message to L2
-  const arbitrator = await hre.ethers.getContractAt('DummyArbitrator', arbitratorAddr, l1Wallet);
-  const adapterParams = ethers.utils.defaultAbiCoder.encode(['uint256'], [gasLimit]);
-  console.log(`The adapter params: ${adapterParams}`);
-  console.log('Prepare to forward the message to L2...');
-  let tx = await arbitrator.forwardMessage(baseL1GatewayAddr, 0, executeCalldata, adapterParams);
-  const txHash = tx.hash;
-  console.log(`The tx hash: ${txHash}`);
-  await tx.wait();
-  console.log(`The transaction has been executed on L1`);
-
-  /**
-   * Query the message informations on L1 via txHash.
-   */
-  const message = (await messenger.getMessagesByTransaction(txHash)).pop();
-  // Waiting for the official base bridge to forward the message to L2
-  const rec = await messenger.waitForMessageReceipt(message);
-  console.log(`The tx receipt: ${JSON.stringify(rec, null, 2)}`);
-  console.log('Done');
-
-  // Example txs:
-  // https://sepolia.etherscan.io/tx/0x00524d9723521e7459581e34013e9a28b5b6d8c4566c3e0b23b2f5fa1726741a
-  // https://sepolia.basescan.org/tx/0xcca496f9fa90e776e6d8e696f12a67c639e0786dab9c84628e039ad5af22bcf7
-});
